feat: Implemented favourites functionality and screen closes #25
Added a page to show favourited accounts, along with the
accompanying functionality to mark an account as a favourite.
Currently there isn't a way to "unfavourite" an account.
Also needed to add a migration to make the functionality work.

Refactored a bit of the code where accounts were being mapped.